Renamed pages and components
