+++ src/pages/QuestionAnswerViewer.jsx
@@ -1,23 +1,23 @@
 import React, { useState, useContext, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
-import UserContext from '@/user';
+
 import Box from '@material-ui/core/Box';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
-import Loading from '@/components/loading/Loading';
 import FormControl from '@material-ui/core/FormControl';
 import InputLabel from '@material-ui/core/InputLabel';
 import MenuItem from '@material-ui/core/MenuItem';
+import Select from '@material-ui/core/Select';
 
 import EditIcon from '@material-ui/icons/Edit';
 import VisibilityIcon from '@material-ui/icons/Visibility';
 
-import Select from '@material-ui/core/Select';
-
+import UserContext from '@/user';
 import API from '@/API';
 import { formatDateTimeNicely } from '@/utils/cache';
 
-import StoredAnswersetViewer from './subComponents/StoredAnswersetViewer';
+import Loading from '@/components/loading/Loading';
+import StoredAnswersetViewer from '@/components/shared/answersetView/StoredAnswersetViewer';
 
 export default function QuestionAnswerViewer() {
   const [question, updateQuestion] = useState(null);
